Python package for the tool (can now be installed with pip)
