+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include excel2wisxml/excel2wisxml_template.xml
