+++ excel2wisxml/excel2wisxmlutils.py
