+++ readme.txt
@@ -5,13 +5,11 @@
 # Version of pyhton
 The script has been tested with python 2.6.6
 
-# Librairies
-4 python libraries must be installed to run the script:
-- xlrd 0.9.4
-- argparse 7.1.2
+# Install the script
+pip install excel2wisxml.tar.gz
 
 # Run the script
-./excel2wisxml.py Metadata-guide-record.xls
+excel2wisxml.py Metadata-guide-record.xls
 where Metadata-guide-record.xls is the excel file containing metadata information
 The script needs the file excel2wisxml_template.xml and excel2wisxmlutils.py to run (in the same directory).
